Adaugare Chest + Usi

+++ src/main/java/mygame/CollisionChecker.java
@@ -9,12 +9,12 @@ public class CollisionChecker {
         this.gp = gp;
     }
 
-    // VERIFICĂ COLIZIUNEA CU TILE-URI (Pereți, Apă etc.)
     public void checkTile(Entity entity) {
-        int entityLeftWorldX = entity.worldX + entity.solidArea.x;
-        int entityRightWorldX = entity.worldX + entity.solidArea.x + entity.solidArea.width;
-        int entityTopWorldY = entity.worldY + entity.solidArea.y;
-        int entityBottomWorldY = entity.worldY + entity.solidArea.y + entity.solidArea.height;
+        // Folosim (int) cast pentru că worldX/worldY sunt double acum
+        int entityLeftWorldX = (int) (entity.worldX + entity.solidArea.x);
+        int entityRightWorldX = (int) (entity.worldX + entity.solidArea.x + entity.solidArea.width);
+        int entityTopWorldY = (int) (entity.worldY + entity.solidArea.y);
+        int entityBottomWorldY = (int) (entity.worldY + entity.solidArea.y + entity.solidArea.height);
 
         int entityLeftCol = entityLeftWorldX / gp.tileSize;
         int entityRightCol = entityRightWorldX / gp.tileSize;
@@ -25,7 +25,8 @@ public class CollisionChecker {
 
         switch (entity.direction) {
             case "up":
-                entityTopRow = (entityTopWorldY - entity.speed) / gp.tileSize;
+                // Scădem viteza (double) și transformăm rezultatul în int pentru a afla rândul
+                entityTopRow = (int) (entityTopWorldY - entity.speed) / gp.tileSize;
                 tileNum1 = gp.tileM.mapTileNum[entityLeftCol][entityTopRow];
                 tileNum2 = gp.tileM.mapTileNum[entityRightCol][entityTopRow];
                 if (gp.tileM.tile[tileNum1].collision || gp.tileM.tile[tileNum2].collision) {
@@ -33,7 +34,7 @@ public class CollisionChecker {
                 }
                 break;
             case "down":
-                entityBottomRow = (entityBottomWorldY + entity.speed) / gp.tileSize;
+                entityBottomRow = (int) (entityBottomWorldY + entity.speed) / gp.tileSize;
                 tileNum1 = gp.tileM.mapTileNum[entityLeftCol][entityBottomRow];
                 tileNum2 = gp.tileM.mapTileNum[entityRightCol][entityBottomRow];
                 if (gp.tileM.tile[tileNum1].collision || gp.tileM.tile[tileNum2].collision) {
@@ -41,7 +42,7 @@ public class CollisionChecker {
                 }
                 break;
             case "left":
-                entityLeftCol = (entityLeftWorldX - entity.speed) / gp.tileSize;
+                entityLeftCol = (int) (entityLeftWorldX - entity.speed) / gp.tileSize;
                 tileNum1 = gp.tileM.mapTileNum[entityLeftCol][entityTopRow];
                 tileNum2 = gp.tileM.mapTileNum[entityLeftCol][entityBottomRow];
                 if (gp.tileM.tile[tileNum1].collision || gp.tileM.tile[tileNum2].collision) {
@@ -49,7 +50,7 @@ public class CollisionChecker {
                 }
                 break;
             case "right":
-                entityRightCol = (entityRightWorldX + entity.speed) / gp.tileSize;
+                entityRightCol = (int) (entityRightWorldX + entity.speed) / gp.tileSize;
                 tileNum1 = gp.tileM.mapTileNum[entityRightCol][entityTopRow];
                 tileNum2 = gp.tileM.mapTileNum[entityRightCol][entityBottomRow];
                 if (gp.tileM.tile[tileNum1].collision || gp.tileM.tile[tileNum2].collision) {
@@ -59,25 +60,23 @@ public class CollisionChecker {
         }
     }
 
-    // VERIFICĂ COLIZIUNEA CU OBIECTE (Chei, Uși etc.)
     public int checkObject(Entity entity, boolean player) {
         int index = 999;
 
         for (int i = 0; i < gp.obj.length; i++) {
             if (gp.obj[i] != null) {
-                // Obținem poziția hitbox-ului entității în lume
-                entity.solidArea.x = entity.worldX + entity.solidArea.x;
-                entity.solidArea.y = entity.worldY + entity.solidArea.y;
+                // Modificăm temporar solidArea folosind cast la int
+                entity.solidArea.x = (int) entity.worldX + entity.solidArea.x;
+                entity.solidArea.y = (int) entity.worldY + entity.solidArea.y;
 
-                // Obținem poziția hitbox-ului obiectului în lume
-                gp.obj[i].solidArea.x = gp.obj[i].worldX + gp.obj[i].solidArea.x;
-                gp.obj[i].solidArea.y = gp.obj[i].worldY + gp.obj[i].solidArea.y;
+                gp.obj[i].solidArea.x = (int) gp.obj[i].worldX + gp.obj[i].solidArea.x;
+                gp.obj[i].solidArea.y = (int) gp.obj[i].worldY + gp.obj[i].solidArea.y;
 
                 switch (entity.direction) {
-                    case "up": entity.solidArea.y -= entity.speed; break;
-                    case "down": entity.solidArea.y += entity.speed; break;
-                    case "left": entity.solidArea.x -= entity.speed; break;
-                    case "right": entity.solidArea.x += entity.speed; break;
+                    case "up": entity.solidArea.y -= (int) entity.speed; break;
+                    case "down": entity.solidArea.y += (int) entity.speed; break;
+                    case "left": entity.solidArea.x -= (int) entity.speed; break;
+                    case "right": entity.solidArea.x += (int) entity.speed; break;
                 }
 
                 if (entity.solidArea.intersects(gp.obj[i].solidArea)) {
@@ -89,10 +88,10 @@ public class CollisionChecker {
                     }
                 }
 
-                // RESETĂM coordonatele hitbox-ului (altfel "pleacă" de pe caracter)
-                entity.solidArea.x = 8; // Valoarea ta default din Player.java
+                // RESETĂM valorile la offset-ul original (int)
+                entity.solidArea.x = 8;
                 entity.solidArea.y = 16;
-                gp.obj[i].solidArea.x = 0; // Valoarea default din SuperObject.java
+                gp.obj[i].solidArea.x = 0;
                 gp.obj[i].solidArea.y = 0;
             }
         }
